Refactored classes to be in separate packages
